+++ GliderProperties.proj
