+++ main.lua
@@ -1,4 +1,4 @@
----------------------------------------------------------------------------------
+require "CiderDebugger";---------------------------------------------------------------------------------
 --
 -- main.lua
 --
